moved get_client() function from cli to storj package. moved read_config() function from cli to storj package.
