+++ storj/__init__.py
@@ -2,16 +2,31 @@
 """Storj package."""
 
 import io
-
+from hashlib import sha256
 
 from ecdsa import SigningKey, SECP256k1
-from hashlib import sha256
+from storj import Client
+from storj.cli import APP_NAME
+from storj.configuration import read_config
 
 from .api import ecdsa_to_hex
 from .http import Client
+from .metadata import __version__
 from .model import Bucket, File, Token
 
-from .metadata import __version__
+
+CFG_EMAIL = 'storj.email'
+CFG_PASSWORD = 'storj.password'
+
+
+def get_client():
+    """Returns a pre-configured Storj HTTP client.
+
+    Returns:
+        (:py:class:`storj.http.Client`): Storj HTTP client.
+    """
+    cfg = read_config()
+    return Client(cfg[CFG_EMAIL], cfg[CFG_PASSWORD])
 
 
 def generate_new_key_pair():
